sstable: split VirtualReader out of reader.go
